+++ C/primitive/elements/env.c
@@ -3,7 +3,7 @@
 #include <stdalign.h>
 #include <stddef.h>
 #include <string.h>
-#include "primitive.h"
+#include "txEnv.h"
 #include "ops.h"
 #include "../../rsort.h"
 #include "../../sha256.h"
@@ -650,29 +650,4 @@ extern tapEnv* simplicity_elements_mallocTapEnv(const rawTapEnv* rawEnv) {
  */
 extern void simplicity_elements_freeTapEnv(tapEnv* env) {
   simplicity_free(env);
-}
-
-/* Construct a txEnv structure from its components.
- * This function will precompute any cached values.
- *
- * Precondition: NULL != tx
- *               NULL != taproot
- *               NULL != genesisHash
- *               ix < tx->numInputs
- */
-txEnv simplicity_build_txEnv(const transaction* tx, const tapEnv* taproot, const sha256_midstate* genesisHash, uint_fast32_t ix) {
-  txEnv result = { .tx = tx
-                 , .taproot = taproot
-                 , .genesisHash = *genesisHash
-                 , .ix = ix
-                 };
-  sha256_context ctx = sha256_init(result.sigAllHash.s);
-  sha256_hash(&ctx, genesisHash);
-  sha256_hash(&ctx, genesisHash);
-  sha256_hash(&ctx, &tx->txHash);
-  sha256_hash(&ctx, &taproot->tapEnvHash);
-  sha256_u32be(&ctx, ix);
-  sha256_finalize(&ctx);
-
-  return result;
 }
